# change aclrtSynchronizeDevice to aclrtSynchronizeStream for better performance